requirements.txt
azure-ai-formrecognizer
python-dotenv
azure-storage-blob
spacy